requirements.txt
boto3
pandas
scipy
scikit-learn
flask
numpy
gunicorn
pytest
openai
requests
pytest-cov
psycopg2-binary
SQLAlchemy
alembic
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
PyJWT